server-test/test/app/requirements.txt
flask
flask-cors
mysql-connector
joblib
